+++ def/ioredis.d.ts
@@ -1,9 +1,7 @@
 /// <reference path="./redis.d.ts" />
 
 declare class A {
-    state: string;
     constructor(opts ?: any, cb ?: any);
-    update(): void;
 }
 
 declare module 'ioredis' {