+++ src/def/ILocalProtocolUser.d.ts
@@ -1,0 +1,8 @@
+interface ILocalProtocolUser {
+    attempts : ILocalProtocolUserAttempt[];
+}
+
+interface ILocalProtocolUserAttempt {
+    account : string;
+    date : number;
+}