+++ src/protocols/list/Local.ts
@@ -4,38 +4,102 @@
 import Core = require("./../Core");
 import _ = require("lodash");
 
+/**
+ * Local
+ *
+ * @module :: Local
+ * @description	:: Local protocol.
+ */
 class Local extends Core {
 
     /**
-     * Contains a list of accounts that have been used for connection
-     * @type {string[]}
-     * @private
+     * ****************************************************
+     * ******************* About Account ******************
+     * ****************************************************
      */
-    private _accountList : string[] = [];
 
     /**
      * Contains a list of account used for authentication in asc order
+     *
+     * @property _accountAttempts
      * @type {string[]}
      * @private
      */
     private _accountAttempts : string[] = [];
 
     /**
-     * Contain the list of user who have submit login form
+     * Contain the list of account currently locked
+     *
+     * @property _accountJail
      * @type {string[]}
      * @private
      */
-    private _userList : string[] = [];
+    private _accountJail : string[] = [];
+
+    /**
+     * Contain information about account actually banned
+     *
+     * @property _accountJailInfo
+     * @type {any[]}
+     * @private
+     */
+    private _accountJailInfo : string[] = [];
+
+    /**
+     * Contains a list of accounts that have been used for connection
+     *
+     * @property _accountList
+     * @type {string[]}
+     * @private
+     */
+    private _accountList : string[] = [];
+
+    /**
+     * ****************************************************
+     * ******************** About User ********************
+     * ****************************************************
+     */
 
     /**
      * Contains a list of IP address used for authentication in asc order
+     *
+     * @property _userAttempts
      * @type {string[]}
      * @private
      */
     private _userAttempts : string[] = [];
 
     /**
+     * Contain the list of user currently banned
+     *
+     * @property _userJail
+     * @type {string[]}
+     * @private
+     */
+    private _userJail : string[] = [];
+
+    /**
+     * Contain information about user actually banned
+     *
+     * @property _userJailInfo
+     * @type {any[]}
+     * @private
+     */
+    private _userJailInfo : any = {};
+
+    /**
+     * Contain the list of user who have submit login form
+     *
+     * @property _userList
+     * @type {ILocalProtocolUser[]}
+     * @private
+     */
+    private _userList : ILocalProtocolUser[] = [];
+
+    /**
      * Contain the list of user IP with a permanent access
+     *
+     * @property _IPWhiteList
      * @type {string[]}
      * @private
      */
@@ -43,12 +107,26 @@ class Local extends Core {
 
     /**
      * Basic constructor
+     *
+     * @param jailConfig {IJailConfig}      Jail configuration
      */
-    public constructor() {
-        super();
+    public constructor(jailConfig : IJailConfig) {
+        super(jailConfig);
 
         // Init name of protocol
         this._protocolName = 'local';
+    }
+
+    /**
+     * Check if account IP is locked
+     * @method _accountIsLocked
+     * @private
+     *
+     * @param accountID {string}    AccountID
+     * @return {boolean}            Return true if account is locked else false
+     */
+    private _accountIsLocked(accountID : string) : boolean {
+        return _.indexOf(this._accountJail, accountID) != -1;
     }
 
     /**
@@ -58,11 +136,8 @@ class Local extends Core {
      * @method _clearQueue
      */
     private _clearQueue() : void {
-        // Debug
-        console.log('Clear queue');
-
         // Calculate current find time of user
-        var userFindTime : number = Date.now() - (this._userFindTime * 1000);
+        var userFindTime : number = Date.now() - this._userFindTime;
 
         // Check last user attempts
         for (var i : number = 0, ls = this._userAttempts.length; i < ls; i++) {
@@ -88,7 +163,7 @@ class Local extends Core {
         }
 
         // Calculate current find time of account
-        var accountFindTime : number = Date.now() - (this._accountFindTime * 1000);
+        var accountFindTime : number = Date.now() - this._accountFindTime;
 
         // Same but with accounts
         for (var i : number = 0, ls = this._accountAttempts.length; i < ls; i++) {
@@ -112,6 +187,26 @@ class Local extends Core {
                 break;
             }
         }
+
+        // Unban user
+        for (var i : number = 0, ls : number = this._userJail.length; i < ls; i++) {
+            if (this._userJailInfo[this._userJail[i]].endBan < Date.now()) {
+                delete (this._userJailInfo[this._userJail[i]]);
+                this._userJail.splice(i, 1);
+                i--;
+                ls--;
+            }
+        }
+
+        // Unlock account
+        for (var i : number = 0, ls : number = this._accountJail.length; i < ls; i++) {
+            if (this._accountJailInfo[this._accountJail[i]].endLock < Date.now()) {
+                delete (this._accountJailInfo[this._accountJail[i]]);
+                this._accountJail.splice(i, 1);
+                i--;
+                ls--;
+            }
+        }
     }
 
     /**
@@ -127,10 +222,111 @@ class Local extends Core {
     }
 
     /**
+     * Check if user IP is banned
+     * @method _userIsBanned
+     * @private
+     *
+     * @param ip {string}       IP address of user
+     * @return {boolean}        Return true if user is banned else false
+     */
+    private _userIsBanned(ip : string) : boolean {
+        return _.indexOf(this._userJail, ip) != -1;
+    }
+
+    /**
+     * Add a new connection attempt
+     * @method addAttempt
+     *
+     * @param accountID {string}    Account ID for auth
+     * @param userIP {string}       User IP for auth
+     */
+    public addAttempt(accountID : string, userIP : string) : void {
+        super.addAttempt(accountID, userIP);
+
+        // Check if user is not already banned and account locked
+        if (this._userIsBanned(userIP) || (this._accountLockEnable && this._accountIsLocked(accountID))) {
+            return;
+        }
+
+        /**
+         * ***************************************************
+         * **************** Add user attempt *****************
+         * ***************************************************
+         */
+
+        // Check if user IP is in list
+        if (!this._userList[userIP]) {
+            this._userList[userIP] = {
+                banTimestamp : 0,
+                attempts : []
+            };
+        }
+
+        // Add attempt
+        this._userList[userIP].attempts.push({
+            account : accountID,
+            date : Date.now()
+        });
+
+        // Add new user attempt
+        this._userAttempts.push(userIP);
+
+        // Check if user attempt is not > max retry
+        if (this._userList[userIP].attempts.length >= this._userMaxRetry) {
+
+            // Send user into jail
+            this._userJail.push(userIP);
+
+            // Store info about end of ban
+            this._userJailInfo[userIP] = {
+                endBan : Date.now() + this._userBanTime
+            }
+        }
+
+        /**
+         * ***************************************************
+         * ************** Add account attempt ****************
+         * ***************************************************
+         */
+
+        // Check if account locker is currently enable
+        if (this._accountLockEnable) {
+
+            // Check if account is in list
+            if (!this._accountList[accountID]) {
+                this._accountList[accountID] = {
+                    lockTimestamp: 0,
+                    attempts: []
+                };
+            }
+
+            // Add account attempt
+            this._accountList[accountID].attempts.push({
+                ip: userIP,
+                date: Date.now()
+            });
+
+            // Add new account attempt
+            this._accountAttempts.push(accountID);
+
+            // Check if account attempt is not > max retry
+            if (this._accountList[accountID].attempts.length >= this._accountMaxRetry) {
+                // Send user into jail
+                this._accountJail.push(accountID);
+
+                // Store info about end of ban
+                this._accountJailInfo[accountID] = {
+                    endLock : Date.now() + this._accountLockTime
+                }
+            }
+        }
+    }
+
+    /**
      * Add a new IP address in white list
      * @method allowIP
      *
-     * @param ip {string}       IP adress of user allowed
+     * @param ip {string}       IP address of user allowed
      * @return {boolean}        Return true if IP is added or false is she's already in DB
      */
     public allowIP(ip : string) : boolean {
@@ -143,17 +339,42 @@ class Local extends Core {
     }
 
     /**
-     * Init protocol
+     * Ban user manually
+     * @method banUser
      *
-     * @method boot
+     * @param ip {string}       IP address of user
+     * @param time {number}     Optional ban time in seconds
      */
-    public boot(protocolConfig : IRedisProtocolConfig, cb : Function) : void {
+    public banUser(ip : string, time ?: number) : void {
+        // Calculate new time of ban
+        var customTime : number = (time) ? time * 1000 : this._userBanTime;
 
+        // Check if user is already banned
+        if (this._userIsBanned(ip)) {
+            // Add more ban time
+            this._userJailInfo[ip].endBan = Date.now() + customTime;
+        } else {
+            // Send user into jail
+            this._userJail.push(ip);
+
+            // Store info about end of ban
+            this._userJailInfo[ip] = {
+                endBan : Date.now() + customTime
+            }
+        }
     }
 
+    /**
+     * Check if user/account are allowed to auth
+     * @method loginAttempt
+     *
+     * @param accountID {string}    Account ID for auth
+     * @param userIP {string}       User IP for auth
+     * @param cb {function}         Callback
+     */
     public loginAttempt(accountID : string, userIP : string, cb : (err ?: any) => void) : void {
         // Debug
-        console.log('Trapjs :: Protocol local :: Login attempt');
+        //console.log('Trapjs :: Protocol local :: Login attempt');
 
         // Clear queue
         this._clearQueue();
@@ -161,45 +382,39 @@ class Local extends Core {
         // Check if user is allowed to go next immediatly
         if (this._userIsAllowed(userIP)) {
             cb();
+        }
+
+        // Check if user is banned and get back error
+        else if (this._userIsBanned(userIP)) {
+            // Send callback with error
+            cb(
+                {
+                    code : 'E_USER_BAN',
+                    user : {
+                        banTime : (this._userJailInfo[userIP].endBan - Date.now()) / 1000
+                    }
+                }
+            );
         } else {
-            // Check if user IP is in list
-            if (_.indexOf(this._userList, userIP) == -1) {
-                //_.findKey(this._userList, 'ip', userIP);
-
-                this._userList[userIP] = {
-                    banTimestamp : 0,
-                    attempts : []
-                };
-            }
-
-            // Add attempt
-            this._userList[userIP].attempts.push({
-                account : accountID,
-                date : Date.now()
-            });
-
-            // Add new user attempt
-            this._userAttempts.push(userIP);
-
-            // Check if account is in list
-            if (_.indexOf(this._accountList, accountID) == -1) {
-                this._accountList[accountID] = {
-                    lockTimestamp : 0,
-                    attempts : []
-                };
-            }
-
-            // Add account attempt
-            this._accountList[accountID].attempts.push({
-                ip : userIP,
-                date : Date.now()
-            });
-
-            // Add new account attempt
-            this._accountAttempts.push(accountID);
-
             // Send callback
             cb();
+        }
+    }
+
+    /**
+     * Unban user manually
+     * @method unbanUser
+     *
+     * @param ip {string}       IP of user who must be unbanned
+     */
+    public unbanUser(ip : string) : void {
+        // Get index of user
+        var userIndex : number = _.indexOf(this._userJail, ip);
+
+        // Check if user in really in jail or not
+        if (userIndex != -1) {
+            delete (this._userJailInfo[ip]);
+            this._userJail.splice(userIndex, 1);
         }
     }
 }