+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Change Log
 
+### v0.0.5 (2015/05/10)
+- Implement Local protocol
+- Begin functional/unit test with mocha
+
 ### v0.0.1 (2015/05/09)
 - Init project